# [Client] Improve post detail page