--- a/packages/api/client/src/hooks/api/post/useGetPostDetail.ts
+++ b/packages/api/client/src/hooks/api/post/useGetPostDetail.ts
@@ -13,5 +13,5 @@ export const useGetPostDetail = (
   useQuery<PostDetailType>(
     postQueryKeys.getPostDetail(seq),
     () => get(postUrl.postDetail(seq)),
-    options
+    { staleTime: 1000, ...options }
   );