--- a/packages/common/src/utils/descriptionFormatting.ts
+++ b/packages/common/src/utils/descriptionFormatting.ts
@@ -1,0 +1,4 @@
+const descriptionFormatting = (description: string) =>
+  description.replace(/\n/g, ' ').replace(/\s+/g, ' ').slice(0, 120);
+
+export default descriptionFormatting;