--- a/packages/common/src/utils/index.ts
+++ b/packages/common/src/utils/index.ts
@@ -3,3 +3,4 @@ export { default as formatDate } from './formatDate';
 export { default as secondsToMs } from './secondsToMs';
 export { default as pxToRem } from './pxToRem';
 export { default as filterImages } from './filterImages';
+export { default as descriptionFormatting } from './descriptionFormatting';